implement basic methods of v2.Client
