+++ v2/doc.go
@@ -1,0 +1,2 @@
+// Package mackerel v2 is now experimental. It is incomplete and unsupported.
+package mackerel
